test: fix for cli tests

+++ tests/cli/test_host.py
@@ -2,25 +2,38 @@ import logging
 
 import pytest
 
-from quads.server import app
 from quads.server.dao.cloud import CloudDao
 from quads.server.dao.host import HostDao
-from tests.cli.config import RESPONSE_DEF_HOST, RESPONSE_RM, HOST, CLOUD, RESPONSE_LS
+from tests.cli.config import HOST, CLOUD, RESPONSE_LS
 from tests.cli.test_base import TestBase
 
 
+def finalizer():
+    host = HostDao.get_host(HOST)
+    if host:
+        HostDao.remove_host(name=HOST)
+    cloud = CloudDao.get_cloud(CLOUD)
+    if cloud:
+        CloudDao.remove_cloud(CLOUD)
+
+
+@pytest.fixture
+def define_fixture(request):
+    request.addfinalizer(finalizer)
+
+    CloudDao.create_cloud(CLOUD)
+
+
+@pytest.fixture
+def remove_fixture(request):
+    request.addfinalizer(finalizer)
+
+    CloudDao.create_cloud(CLOUD)
+    HostDao.create_host(HOST, "r640", "scalelab", CLOUD)
+
+
 class TestHost(TestBase):
-
-    def test_define_host(self):
-        def setup():
-            CloudDao.create_cloud(CLOUD)
-
-        def teardown():
-            HostDao.remove_host(name=HOST)
-            CloudDao.remove_cloud(CLOUD)
-
-        setup()
-
+    def test_define_host(self, define_fixture):
         self.cli_args["hostresource"] = HOST
         self.cli_args["hostcloud"] = CLOUD
         self.cli_args["hosttype"] = "scalelab"
@@ -32,20 +45,7 @@ class TestHost(TestBase):
         assert host is not None
         assert host.name == HOST
 
-        # teardown()
-
-    def test_ls_host(self):
-        with self._caplog.at_level(logging.INFO, logger="test_log"):
-            self.quads_cli_call("ls_hosts")
-        assert self._caplog.messages[0] == RESPONSE_LS
-
-    def test_remove_host(self, ):
-        def setup():
-            CloudDao.create_cloud(CLOUD)
-            HostDao.create_host(HOST, "r640", "scalelab", CLOUD)
-
-        setup()
-
+    def test_remove_host(self, remove_fixture):
         self.cli_args["host"] = HOST
 
         host = HostDao.get_host(HOST)
@@ -57,4 +57,3 @@ class TestHost(TestBase):
 
         host = HostDao.get_host(HOST)
         assert not host
-
